requirements.txt
black==22.3.0
boto3==1.22.12
botocore==1.25.12
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
confluent-kafka==1.8.2
idna==3.3
jmespath==1.0.0
kafka-python==2.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-dateutil==2.8.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
